# Faster shard scaling