--- a/quickwit/quickwit-config/src/cluster_config/mod.rs
+++ b/quickwit/quickwit-config/src/cluster_config/mod.rs
@@ -24,7 +24,7 @@ pub struct ClusterConfig {
     pub default_index_root_uri: Uri,
     pub replication_factor: usize,
     pub shard_throughput_limit: ByteSize,
-    pub shard_scaling_factor: f32,
+    pub shard_scale_up_factor: f32,
 }
 
 impl ClusterConfig {
@@ -36,7 +36,7 @@ impl ClusterConfig {
             default_index_root_uri: Uri::for_test("ram:///indexes"),
             replication_factor: 1,
             shard_throughput_limit: quickwit_common::shared_consts::DEFAULT_SHARD_THROUGHPUT_LIMIT,
-            shard_scaling_factor: 1.01,
+            shard_scale_up_factor: 1.01,
         }
     }
 }