… quickwit/quickwit-control-plane/src/ingest/scaling_arbiter.rs