### Description

Indexers return a lot of 429 when ingest rates fluctuate, even if the overall indexing capacity (~5MB/core) is note reached. This PR makes some small parameter adjustments to decrease the overall error rates on typical workloads. It will be the topic for another issue/PR to introduce the structual changes that are necessary to improve the routing strategy.

- Customizable shard burst limit with default increased to 50MB: helps absorbing the load until the next scale up
- Shard scaling factor: helps keeping up with load fluctuations on workloads with high number of shards (e.g having +20MB/s on an index where the average ingest rate is 100MB/s is more likely than +20MB/s on an index where the average rate is 1MB/s)

Related to https://github.com/quickwit-oss/quickwit/issues/5270

### How was this PR tested?

Adhoc python load testing framework: https://github.com/rdettai/qw-ingest-tests