@rdettai 2025-02-13
Observations so far:
- there is a high latency between a scaling action and the entry being added to the router (5+s)
- even with a high `shard_burst_limit` (100MB), a fairly low ingest rate (20MB/s) runs into throttling before the shards are scaled an the routers know about the new shards
- TODO: check when the extra shard is taken into account for the average shard throughput calculation
  - if it is already accounted for while it isn't known by routers, that's very bad because it means the arbiter will think the load has lowered when we are actually rejecting some ingest that could have gone to the un-registered shards
  - ideally, it should be only accounted for once it has been broadcasted (even though that wouldn't work well if we add other shard knowledge propagation techniques in the future)